==> 8-User_Interface.ipynb <==
# Welcome to MovieRec4Parents!

#### Please do not tamper with the cell below, or the interface won't work.
#### *** IMPORTANT: ***  If you forked and cloned this from my GitHub, you will need to contact also download the cosine similarity matrix file, similarity_matrix_tfidfvec_truncSVD4000.pkl, for this recommender to work. You can do so by going to https://drive.google.com/open?id=1MZv3t0YlG6VIQYRr5c0_e88stkJFFO1z and downloading the file. Be sure to move it to the  /capstone/data folder before running any of the cells below.


from lib.interface import *

### If you want five quick movie recommendations, think of a movie that is like a movie you'd like to see, click next to "find_similar_movies" and press enter. If you have more time and don't mind answering a few questions, click next to "get_parent_ratings_first" and press enter. You'll be asked a few questions which will be used to refine my recommendations.

find_similar_movies()

get_parent_ratings_first()

### If you'd like more recs, rerun the cell(s) above! And thank you for using MovieRec4Parents!
